added retry logic to kubectl and helm plugin (#18)
To prevent the whole setup from failing when the apiserver is temporarily unavailable, the kubectl as well as the helm plugin now try several times, if the command doesn't succeed.

The default is 5 attempts while waiting 15 seconds in between. These values can be configured via the following block in the plugin's values:

```yaml
retry:
  attempts: 5
  wait: 15
```

If something gets deleted via kubectl because it was there in a previous deployment and isn't there in a new deployment, there is no place to specify these values, so in this case, the defaults will be used.