add possibility to define an installation handler
The acre.yaml can now define a path to an installation handler script.
To do so, use this in your acre.yaml:
```yaml
meta:
  installationHandler: <path-to-executable-bash-file>
```

If a script is given, it will be called once at the beginning (with argument `prepare`) and once at the end (with argument `finalize`) of each sow command (except for `sow version` and `sow help`).