improve git plugin
The git plugin now saves which tag/branch/commit it checks out.
If it is called to check out the same version, it first checks whether all given files to check out are there and if so, it doesn't do anything.
If another version is to be checked out or a file is not found, the repository is checked out anew.

The git plugin now also checks out files in delete mode, as they might be needed for another plugin.